Version: Use cmakedefine01 for boolean
The CMake_VERSION_IS_DIRTY CMake variable is not set by
CMakeVersionCompute or its delegate if not building from git, or if the
git executable is not found.

That means that cmVersionConfig.h would be generated to contain an empty
C++ define for CMake_VERSION_IS_DIRTY.

cmake::ReportCapabilities() requires the preprocessor define to have a
value, so use the standard construct for booleans of cmakedefine01.