Add SOURCE_SUBDIR option to ExternalProject
Add a new SOURCE_SUBDIR option to ExternalProject_Add that allows
specifying the location of the CMakeLists.txt to use as the project root
relative to the SOURCE_DIR.

This is helpful for projects that have odd layouts, or projects that
provide both a superbuild and project-only build depending on which
CMakeLists.txt is used.

See also http://stackoverflow.com/questions/30028117.

Fixes #15118.