cmCTest: don't redefine cout and cerr
The definitions have been introduced to ensure that cout and cerr are
not used in certain files.  However, this limitation does not apply to
all source files that require cmCTest.h to be included.  Furthermore,
the definitions cause side effects depending on the include order.  In
total, the definitions do more harm than good.  Remove them.