FindX11: use pkg-config if found
On Linux pkg-config is the de facto standard way of discovering
build-time dependencies and xorg projects have been distributing
pkg-config .pc files for a long time now.

Developers may manage the PKG_CONFIG_PATH environment variable as a way
to select what particular build of a dependency they want to use and it
can be awkward if cmake doesn't honour this automatically (without
requiring the manual use of FindPkgConfig)

The large number of X11 libraries that may be installed to a
non-standard prefix during development are especially awkward to have to
override for cmake based projects.